# worldwide-framework/openapi-worldwide/src/main/resources/openapi/json/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "全球领域模型",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/product/category": {
      "post": {
        "summary": "CreateCategory",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "创建商品分类",
        "operationId": "CreateCategory",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCategoryReq",
                "x-apifox-overrides": {},
                "title": "创建商品分类请求DTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "title": "商品分类编码",
                  "type": "string",
                  "pattern": "^\\d+$"
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/product/category/{id}": {
      "delete": {
        "summary": "DeleteCategory",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "删除商品分类",
        "operationId": "DeleteCategory",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "商品分类编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "是否删除成功"
                },
                "examples": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "UpdateCategory",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "更新商品分类",
        "operationId": "UpdateCategory",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "商品分类编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCategoryReq",
                "x-apifox-overrides": {},
                "title": "更新商品分类请求DTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "是否更新成功"
                },
                "examples": {}
              }
            }
          }
        }
      },
      "get": {
        "summary": "GetCategory",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "获取商品分类",
        "operationId": "GetCategory",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "商品分类编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategoryRsp",
                  "x-apifox-overrides": {},
                  "title": "商品分类响应"
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProductCategoryReq": {
        "type": "object",
        "properties": {
          "parentId": {
            "title": "父分类编码",
            "type": "string",
            "pattern": "^\\d+$"
          },
          "name": {
            "type": "string",
            "title": "分类名称"
          },
          "icon": {
            "title": "分类图标",
            "type": "string"
          },
          "bannerUrl": {
            "type": "string",
            "title": "分类图片链接"
          },
          "sort": {
            "type": "integer",
            "title": "排序"
          },
          "description": {
            "title": "分类描述",
            "type": "string"
          },
          "status": {
            "type": "integer",
            "title": "开启状态",
            "enum": [
              0,
              1
            ],
            "x-apifox": {
              "enumDescriptions": {
                "0": "关闭",
                "1": "开启"
              }
            }
          }
        },
        "x-apifox-orders": [
          "parentId",
          "name",
          "icon",
          "bannerUrl",
          "sort",
          "description",
          "status"
        ],
        "required": [
          "parentId",
          "name",
          "bannerUrl",
          "sort",
          "status"
        ],
        "title": "创建商品分类请求DTO",
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
...
      "ProductCategoryRsp": {
        "type": "object",
        "properties": {
          "id": {
            "title": "分类编码",
            "type": "string",
            "pattern": "^\\d+$"
          },
          "parentId": {
            "title": "父分类编码",
            "type": "string",
            "pattern": "^\\d+$"
          },
          "name": {
            "type": "string",
            "title": "分类名称"
          },
          "icon": {
            "type": "string",
            "title": "分类图标"
          },
          "bannerUrl": {
            "type": "string",
            "title": "分类图片链接"
          },
          "sort": {
            "type": "integer",
            "title": "排序"
          },
          "description": {
            "type": "string",
            "title": "分类描述"
          },
          "createdTime": {
            "type": "string",
            "title": "创建时间"
          },
          "updatedTime": {
            "type": "string",
            "title": "更新时间"
          },
          "createdBy": {
            "title": "创建者",
            "type": "string",
            "pattern": "^\\d+$"
          },
          "updatedBy": {
            "title": "更新者",
            "type": "string",
            "pattern": "^\\d+$"
          },
          "status": {
            "type": "integer",
            "title": "开启状态",
            "enum": [
              0,
              1
            ],
            "x-apifox": {
              "enumDescriptions": {
                "0": "关闭",
                "1": "开启"
              }
            }
          }
        },
        "x-apifox-orders": [
          "id",
          "parentId",
          "name",
          "icon",
          "bannerUrl",
          "sort",
          "description",
          "createdTime",
          "updatedTime",
          "createdBy",
          "updatedBy",
          "status"
        ],
        "title": "商品分类响应DTO",
        "required": [
          "id",
          "parentId",
          "name",
          "bannerUrl",
          "sort",
          "createdTime",
          "createdBy",
          "status"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
...
      "UpdateProductCategoryReq": {
        "type": "object",
        "properties": {
          "name": {
            "title": "分类名称",
            "type": "string"
          },
          "icon": {
            "title": "分类图标",
            "type": "string"
          },
          "bannerUrl": {
            "title": "分类图片链接",
            "type": "string"
          },
          "sort": {
            "title": "排序",
            "type": "integer"
          },
          "description": {
            "title": "分类描述",
            "type": "string"
          },
          "status": {
            "title": "开启状态",
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "x-apifox": {
              "enumDescriptions": {
                "0": "关闭",
                "1": "开启"
              }
            }
          }
        },
        "x-apifox-orders": [
          "name",
          "icon",
          "bannerUrl",
          "sort",
          "description",
          "status"
        ],
        "title": "更新商品分类请求DTO",
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
...
    }
  }
}